// src/model/Bookmark.ts
export interface Bookmark {
  // Autogenerated ID
  id: string;
  // Name of the bookmark
  name: string;
  // URL of the bookmark
  url: string;
  // Whether or not to show the favicon
  // located at: {URL}/favicon.ico
  useFavicon: boolean;
  // Simple icons slug.
  // See: https://www.npmjs.com/package/simple-icons
  simpleIconsSlug?: string;
}